refactor: SRP split of large feature files into focused single-responsibility modules
mirror/:
- extract MirrorControlPanel.kt (animated control pill composable)
- extract MirrorCoordinateTransform.kt (pure projectCoordinates() geometry helper)
- trim MirrorScreen.kt (~858 → ~550 lines)

keyboard/:
- extract KeyboardKeyCap.kt (KeyBounds, findKeyInLayout(), KeyCap composable)
- extract KeyboardMouseOverlay.kt (mouse button overlay composables)
- trim KeyboardScreen.kt (~900 → ~500 lines)

macropad/:
- extract PadActionPicker.kt (ActionCategory, ActionPicker, sub-pickers, KEYBOARD_KEY_PRESETS)
- extract PadButtonEditDialog.kt (ButtonEditDialog, SectionLabel)
- extract PadCanvas.kt (PadCanvas, DraggableButton)
- extract MacroPadButton.kt (PadButton, ScrollWheelFace, button constants)
- extract MacroPadActionDispatch.kt (injectActionDown/Up helpers)
- trim MacroPadEditor.kt (~1500 → 661 lines)
- trim MacroPadScreen.kt (501 → 322 lines)

settings/:
- extract ToolSettingsComponents.kt (RememberSettingRow, dropdowns, enum label helpers)
- extract MirrorToolSettings.kt (mirror-mode settings composable)
- extract KeyboardToolSettings.kt (keyboard-mode settings composable)
- extract TouchpadToolSettings.kt (touchpad-mode settings composable)
- extract GlobalSettingsComponents.kt (all row composables for GlobalSettingsScreen)
- trim ToolSettingsPanel.kt (612 → 224 lines, orchestrator only)
- trim GlobalSettingsScreen.kt (492 → 194 lines, scaffold + state only)

docs: update AGENTS.md §3 package structure with all 16 new files
