+++ app/src/main/java/com/stormpanda/megingiard/mirror/MirrorCoordinateTransform.kt
@@ -1,0 +1,56 @@
+package com.stormpanda.megingiard.mirror
+
+/**
+ * Maps a raw touch position on the mirror surface back through the current zoom/pan
+ * transform to obtain the normalised content coordinate [0, 1] that corresponds to
+ * the touched point on the primary display.
+ *
+ * The SurfaceView is centered in the secondary display's FrameLayout, so its pivot
+ * point for the scale/translate transform lies at the screen center (screenW/2, screenH/2),
+ * NOT at the SurfaceView's own center in its local coordinate space (sw/2, sh/2).
+ * These differ when the content is letterboxed (sw != screenW or sh != screenH).
+ *
+ * Visual transform (screen → SurfaceView local):
+ *   screenPos = screenCenter + (svPos - svCenter) * scale + offset
+ *   svPos     = (screenPos  - screenCenter - offset) / scale + svCenter
+ *
+ * Returns `null` when the touch lands outside the visible content area (e.g. letterbox
+ * bars), in which case the caller should not inject the touch.
+ *
+ * @param touchX   Raw X of the touch on the secondary display (pixels)
+ * @param touchY   Raw Y of the touch on the secondary display (pixels)
+ * @param screenW  Full width of the secondary display Compose surface (gestureBoxSize.width)
+ * @param screenH  Full height of the secondary display Compose surface (gestureBoxSize.height)
+ * @param sw       Width of the letterboxed content area = ScreenCaptureManager.surfaceWidth
+ * @param sh       Height of the letterboxed content area = ScreenCaptureManager.surfaceHeight
+ * @param scale    Current zoom scale (1.0 = no zoom)
+ * @param offsetX  Current pan offset X (pixels)
+ * @param offsetY  Current pan offset Y (pixels)
+ * @return Pair(normalizedX, normalizedY) or null if out-of-bounds
+ */
+internal fun projectCoordinates(
+    touchX: Float,
+    touchY: Float,
+    screenW: Float,
+    screenH: Float,
+    sw: Float,
+    sh: Float,
+    scale: Float,
+    offsetX: Float,
+    offsetY: Float
+): Pair<Float, Float>? {
+    if (sw <= 0f || sh <= 0f || scale <= 0f || screenW <= 0f || screenH <= 0f) return null
+    // Screen-space center — this is where the SurfaceView is anchored (CENTER gravity).
+    val screenCenterX = screenW / 2f
+    val screenCenterY = screenH / 2f
+    // SurfaceView-local pivot for the scale transform.
+    val svCenterX = sw / 2f
+    val svCenterY = sh / 2f
+    // Invert: svPos = (screenPos - screenCenter - offset) / scale + svCenter
+    val svX = (touchX - screenCenterX - offsetX) / scale + svCenterX
+    val svY = (touchY - screenCenterY - offsetY) / scale + svCenterY
+    val nx = svX / sw
+    val ny = svY / sh
+    if (nx !in 0f..1f || ny !in 0f..1f) return null
+    return Pair(nx, ny)
+}
