+++ app/src/main/java/com/stormpanda/megingiard/settings/ToolSettingsPanel.kt
@@ -1,8 +1,8 @@
 package com.stormpanda.megingiard.settings
 
+import androidx.activity.compose.BackHandler
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
@@ -16,20 +16,12 @@ import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.foundation.verticalScroll
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.automirrored.filled.ArrowForward
-import androidx.compose.material.icons.filled.ArrowDropDown
 import androidx.compose.material.icons.filled.Close
 import androidx.compose.material.icons.filled.Settings
-import androidx.compose.material3.Checkbox
-import androidx.compose.material3.CheckboxDefaults
-import androidx.compose.material3.DropdownMenu
-import androidx.compose.material3.DropdownMenuItem
 import androidx.compose.material3.HorizontalDivider
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
-import androidx.compose.material3.Switch
-import androidx.compose.material3.SwitchDefaults
 import androidx.compose.material3.Text
-import androidx.activity.compose.BackHandler
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
@@ -229,384 +221,4 @@ fun ToolSettingsPanel(
     }
 }
 
-@Composable
-private fun MirrorToolSettings(
-    autoStartCapture: Boolean,
-    accentColor: Color,
-    onAutoStartChanged: (Boolean) -> Unit,
-    rememberViewport: Boolean,
-    onRememberViewportChanged: (Boolean) -> Unit,
-    rememberLock: Boolean,
-    onRememberLockChanged: (Boolean) -> Unit,
-    rememberProjection: Boolean,
-    onRememberProjectionChanged: (Boolean) -> Unit,
-    pinchWhileProjecting: Boolean,
-    onPinchWhileProjectingChanged: (Boolean) -> Unit
-) {
-    val colors = LocalAppColors.current
-    Column(verticalArrangement = Arrangement.spacedBy(12.dp)) {
-        // Auto-start capture
-        Row(
-            modifier = Modifier.fillMaxWidth(),
-            verticalAlignment = Alignment.CenterVertically
-        ) {
-            Column(modifier = Modifier.weight(1f)) {
-                Text(
-                    text = stringResource(R.string.settings_auto_start_capture),
-                    color = colors.onSurface,
-                    fontSize = 14.sp
-                )
-                Text(
-                    text = stringResource(R.string.settings_auto_start_capture_desc),
-                    color = colors.onSurfaceSecondary,
-                    fontSize = 12.sp
-                )
-            }
-            Switch(
-                checked = autoStartCapture,
-                onCheckedChange = onAutoStartChanged,
-                colors = SwitchDefaults.colors(
-                    checkedThumbColor = Color.White,
-                    checkedTrackColor = accentColor
-                )
-            )
-        }
 
-        HorizontalDivider(color = colors.divider)
-
-        // Remember viewport
-        RememberSettingRow(
-            label = stringResource(R.string.settings_remember_viewport),
-            description = stringResource(R.string.settings_remember_viewport_desc),
-            checked = rememberViewport,
-            accentColor = accentColor,
-            onCheckedChange = onRememberViewportChanged
-        )
-
-        // Remember lock
-        RememberSettingRow(
-            label = stringResource(R.string.settings_remember_lock),
-            description = stringResource(R.string.settings_remember_lock_desc),
-            checked = rememberLock,
-            accentColor = accentColor,
-            onCheckedChange = onRememberLockChanged
-        )
-
-        // Remember touch projection
-        RememberSettingRow(
-            label = stringResource(R.string.settings_remember_projection),
-            description = stringResource(R.string.settings_remember_projection_desc),
-            checked = rememberProjection,
-            accentColor = accentColor,
-            onCheckedChange = onRememberProjectionChanged
-        )
-
-        HorizontalDivider(color = colors.divider)
-
-        // Pinch-to-zoom while projecting
-        RememberSettingRow(
-            label = stringResource(R.string.settings_pinch_while_projecting),
-            description = stringResource(R.string.settings_pinch_while_projecting_desc),
-            checked = pinchWhileProjecting,
-            accentColor = accentColor,
-            onCheckedChange = onPinchWhileProjectingChanged
-        )
-    }
-}
-
-@Composable
-private fun RememberSettingRow(
-    label: String,
-    description: String,
-    checked: Boolean,
-    accentColor: Color,
-    onCheckedChange: (Boolean) -> Unit
-) {
-    val colors = LocalAppColors.current
-    Row(
-        modifier = Modifier.fillMaxWidth(),
-        verticalAlignment = Alignment.CenterVertically
-    ) {
-        Column(modifier = Modifier.weight(1f)) {
-            Text(
-                text = label,
-                color = colors.onSurface,
-                fontSize = 14.sp
-            )
-            Text(
-                text = description,
-                color = colors.onSurfaceSecondary,
-                fontSize = 12.sp
-            )
-        }
-        Checkbox(
-            checked = checked,
-            onCheckedChange = onCheckedChange,
-            colors = CheckboxDefaults.colors(
-                checkedColor = accentColor,
-                checkmarkColor = Color.White,
-                uncheckedColor = colors.onSurfaceSecondary
-            )
-        )
-    }
-}
-
-@Composable
-private fun KeyboardToolSettings(
-    kbLayout: KbLayout,
-    onKbLayoutChanged: (KbLayout) -> Unit,
-    kbTrackpointEnabled: Boolean,
-    onKbTrackpointEnabledChanged: (Boolean) -> Unit,
-    kbMouseBtnPos: KbMouseBtnPos,
-    onKbMouseBtnPosChanged: (KbMouseBtnPos) -> Unit,
-    kbRepeatEnabled: Boolean,
-    onKbRepeatEnabledChanged: (Boolean) -> Unit,
-    kbFullscreen: Boolean,
-    onKbFullscreenChanged: (Boolean) -> Unit,
-) {
-    val colors = LocalAppColors.current
-    val accentColor = colors.accent
-    Column(verticalArrangement = Arrangement.spacedBy(12.dp)) {
-        LayoutDropdownRow(
-            currentLayout = kbLayout,
-            onLayoutSelected = onKbLayoutChanged,
-            accentColor = accentColor,
-        )
-        RememberSettingRow(
-            label = stringResource(R.string.settings_kb_trackpoint),
-            description = stringResource(R.string.settings_kb_trackpoint_desc),
-            checked = kbTrackpointEnabled,
-            onCheckedChange = onKbTrackpointEnabledChanged,
-            accentColor = accentColor,
-        )
-        if (kbTrackpointEnabled) {
-            MouseBtnPosDropdownRow(
-                currentPos = kbMouseBtnPos,
-                onPosSelected = onKbMouseBtnPosChanged,
-                accentColor = accentColor,
-            )
-        }
-        RememberSettingRow(
-            label = stringResource(R.string.settings_kb_repeat),
-            description = stringResource(R.string.settings_kb_repeat_desc),
-            checked = kbRepeatEnabled,
-            onCheckedChange = onKbRepeatEnabledChanged,
-            accentColor = accentColor,
-        )
-        RememberSettingRow(
-            label = stringResource(R.string.settings_kb_fullscreen),
-            description = stringResource(R.string.settings_kb_fullscreen_desc),
-            checked = kbFullscreen,
-            onCheckedChange = onKbFullscreenChanged,
-            accentColor = accentColor,
-        )
-    }
-}
-
-@Composable
-private fun LayoutDropdownRow(
-    currentLayout: KbLayout,
-    onLayoutSelected: (KbLayout) -> Unit,
-    accentColor: Color,
-) {
-    var expanded by remember { mutableStateOf(false) }
-    val colors = LocalAppColors.current
-    Row(
-        modifier = Modifier.fillMaxWidth(),
-        verticalAlignment = Alignment.CenterVertically
-    ) {
-        Column(modifier = Modifier.weight(1f)) {
-            Text(
-                text = stringResource(R.string.settings_kb_layout),
-                color = colors.onSurface,
-                fontSize = 14.sp
-            )
-            Text(
-                text = stringResource(R.string.settings_kb_layout_desc),
-                color = colors.onSurfaceSecondary,
-                fontSize = 12.sp
-            )
-        }
-        Box {
-            Row(
-                modifier = Modifier
-                    .clickable { expanded = true }
-                    .background(colors.onSurface.copy(alpha = 0.08f), RoundedCornerShape(8.dp))
-                    .padding(horizontal = 12.dp, vertical = 6.dp),
-                verticalAlignment = Alignment.CenterVertically
-            ) {
-                Text(
-                    text = stringResource(currentLayout.labelResId()),
-                    color = colors.onSurface,
-                    fontSize = 14.sp
-                )
-                Icon(
-                    imageVector = Icons.Filled.ArrowDropDown,
-                    contentDescription = null,
-                    tint = colors.onSurfaceSecondary,
-                )
-            }
-            DropdownMenu(
-                expanded = expanded,
-                onDismissRequest = { expanded = false },
-                modifier = Modifier.background(colors.surface)
-            ) {
-                KbLayout.entries.forEach { layout ->
-                    DropdownMenuItem(
-                        text = {
-                            Text(
-                                text = stringResource(layout.labelResId()),
-                                color = if (layout == currentLayout) accentColor else colors.onSurface,
-                                fontSize = 14.sp
-                            )
-                        },
-                        onClick = { onLayoutSelected(layout); expanded = false }
-                    )
-                }
-            }
-        }
-    }
-}
-
-private fun KbLayout.labelResId(): Int = when (this) {
-    KbLayout.QWERTZ -> R.string.settings_kb_layout_qwertz
-    KbLayout.QWERTY -> R.string.settings_kb_layout_qwerty
-    KbLayout.AZERTY -> R.string.settings_kb_layout_azerty
-}
-
-private fun KbMouseBtnPos.labelResId(): Int = when (this) {
-    KbMouseBtnPos.LEFT  -> R.string.kb_mouse_btn_pos_left
-    KbMouseBtnPos.RIGHT -> R.string.kb_mouse_btn_pos_right
-    KbMouseBtnPos.BOTH  -> R.string.kb_mouse_btn_pos_both
-}
-
-@Composable
-private fun MouseBtnPosDropdownRow(
-    currentPos: KbMouseBtnPos,
-    onPosSelected: (KbMouseBtnPos) -> Unit,
-    accentColor: Color,
-) {
-    var expanded by remember { mutableStateOf(false) }
-    val colors = LocalAppColors.current
-    Row(
-        modifier = Modifier.fillMaxWidth(),
-        verticalAlignment = Alignment.CenterVertically
-    ) {
-        Column(modifier = Modifier.weight(1f)) {
-            Text(
-                text = stringResource(R.string.settings_kb_mouse_btn_pos),
-                color = colors.onSurface,
-                fontSize = 14.sp
-            )
-            Text(
-                text = stringResource(R.string.settings_kb_mouse_btn_pos_desc),
-                color = colors.onSurfaceSecondary,
-                fontSize = 12.sp
-            )
-        }
-        Box {
-            Row(
-                modifier = Modifier
-                    .clickable { expanded = true }
-                    .background(colors.onSurface.copy(alpha = 0.08f), RoundedCornerShape(8.dp))
-                    .padding(horizontal = 12.dp, vertical = 6.dp),
-                verticalAlignment = Alignment.CenterVertically
-            ) {
-                Text(
-                    text = stringResource(currentPos.labelResId()),
-                    color = colors.onSurface,
-                    fontSize = 14.sp
-                )
-                Icon(
-                    imageVector = Icons.Filled.ArrowDropDown,
-                    contentDescription = null,
-                    tint = colors.onSurfaceSecondary,
-                )
-            }
-            DropdownMenu(
-                expanded = expanded,
-                onDismissRequest = { expanded = false },
-                modifier = Modifier.background(colors.surface)
-            ) {
-                KbMouseBtnPos.entries.forEach { pos ->
-                    DropdownMenuItem(
-                        text = {
-                            Text(
-                                text = stringResource(pos.labelResId()),
-                                color = if (pos == currentPos) accentColor else colors.onSurface,
-                                fontSize = 14.sp
-                            )
-                        },
-                        onClick = { onPosSelected(pos); expanded = false }
-                    )
-                }
-            }
-        }
-    }
-}
-
-@Composable
-private fun TouchpadToolSettings(
-    touchpadUseMouse: Boolean,
-    onTouchpadUseMouseChanged: (Boolean) -> Unit,
-    tapToClick: Boolean,
-    onTapToClickChanged: (Boolean) -> Unit,
-    twoFingerTap: Boolean,
-    onTwoFingerTapChanged: (Boolean) -> Unit,
-    accentColor: Color,
-) {
-    Column(verticalArrangement = Arrangement.spacedBy(12.dp)) {
-        InputMethodRow(
-            label = stringResource(R.string.settings_input_method),
-            description = stringResource(R.string.settings_input_method_desc),
-            useMouse = touchpadUseMouse,
-            onUseMouseChanged = onTouchpadUseMouseChanged,
-            accentColor = accentColor
-        )
-        if (touchpadUseMouse) {
-            HorizontalDivider(color = LocalAppColors.current.divider)
-            RememberSettingRow(
-                label = stringResource(R.string.settings_tap_to_click),
-                description = stringResource(R.string.settings_tap_to_click_desc),
-                checked = tapToClick,
-                onCheckedChange = onTapToClickChanged,
-                accentColor = accentColor
-            )
-            RememberSettingRow(
-                label = stringResource(R.string.settings_two_finger_tap),
-                description = stringResource(R.string.settings_two_finger_tap_desc),
-                checked = twoFingerTap,
-                onCheckedChange = onTwoFingerTapChanged,
-                accentColor = accentColor
-            )
-        }
-    }
-}
-
-@Composable
-private fun InputMethodRow(
-    label: String,
-    description: String,
-    useMouse: Boolean,
-    onUseMouseChanged: (Boolean) -> Unit,
-    accentColor: Color,
-) {
-    val colors = LocalAppColors.current
-    Row(
-        modifier = Modifier.fillMaxWidth(),
-        verticalAlignment = Alignment.CenterVertically
-    ) {
-        Column(modifier = Modifier.weight(1f)) {
-            Text(text = label, color = colors.onSurface, fontSize = 14.sp)
-            Text(text = description, color = colors.onSurfaceSecondary, fontSize = 12.sp)
-        }
-        Switch(
-            checked = useMouse,
-            onCheckedChange = onUseMouseChanged,
-            colors = SwitchDefaults.colors(
-                checkedThumbColor = Color.White,
-                checkedTrackColor = accentColor
-            )
-        )
-    }
-}
